requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
nemollm==0.3.5
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.335
marshmallow==3.20.1
gunicorn==21.2.0
redis==5.0.1
